+++ exonum-java-binding-core/src/main/java/com/exonum/binding/service/NodeFake.java
@@ -16,9 +16,11 @@
 
 package com.exonum.binding.service;
 
+import static com.exonum.binding.common.crypto.CryptoFunctions.Ed25519.PUBLIC_KEY_BYTES;
 import static com.exonum.binding.common.hash.Hashing.DEFAULT_HASH_SIZE_BYTES;
 import static com.google.common.base.Preconditions.checkNotNull;
 
+import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.proxy.Cleaner;
 import com.exonum.binding.proxy.CloseFailuresException;
@@ -59,7 +61,7 @@ public final class NodeFake implements Node {
 
   private final MemoryDb database;
 
-  private final byte[] publicKey;
+  private final PublicKey publicKey;
 
   /**
    * Creates a new node fake with the given database and an empty public key.
@@ -67,7 +69,7 @@ public final class NodeFake implements Node {
    * @param database a database to provide snapshots of
    */
   public NodeFake(MemoryDb database) {
-    this(database, new byte[0]);
+    this(database, PublicKey.fromBytes(new byte[PUBLIC_KEY_BYTES]));
   }
 
   /**
@@ -76,9 +78,9 @@ public final class NodeFake implements Node {
    * @param database a database to provide snapshots of
    * @param publicKey a public key of the node
    */
-  public NodeFake(MemoryDb database, byte[] publicKey) {
+  public NodeFake(MemoryDb database, PublicKey publicKey) {
     this.database = checkNotNull(database);
-    this.publicKey = publicKey.clone();
+    this.publicKey = publicKey;
   }
 
   /**
@@ -105,8 +107,8 @@ public final class NodeFake implements Node {
   }
 
   @Override
-  public byte[] getPublicKey() {
-    return publicKey.clone();
+  public PublicKey getPublicKey() {
+    return publicKey;
   }
 
   /**