+++ exonum-java-binding-core/src/main/java/com/exonum/binding/service/NodeProxy.java
@@ -16,6 +16,7 @@
 
 package com.exonum.binding.service;
 
+import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.proxy.AbstractCloseableNativeProxy;
 import com.exonum.binding.proxy.Cleaner;
@@ -96,8 +97,9 @@ public final class NodeProxy extends AbstractCloseableNativeProxy implements Nod
    * @throws IllegalStateException if the node proxy is closed
    */
   @Override
-  public byte[] getPublicKey() {
-    return nativeGetPublicKey(getNativeHandle());
+  public PublicKey getPublicKey() {
+    byte[] publicKey = nativeGetPublicKey(getNativeHandle());
+    return PublicKey.fromBytes(publicKey);
   }
 
   private native byte[] nativeGetPublicKey(long nativeHandle);