+++ exonum-java-binding-core/src/main/java/com/exonum/binding/storage/indices/ConfigurableRustIter.java
@@ -18,8 +18,7 @@ package com.exonum.binding.storage.indices;
 
 import com.exonum.binding.proxy.AbstractNativeProxy;
 import com.exonum.binding.proxy.NativeHandle;
-import com.exonum.binding.storage.database.View;
-import com.exonum.binding.storage.database.ViewModificationCounter;
+import com.exonum.binding.storage.database.ModificationCounter;
 import java.util.ConcurrentModificationException;
 import java.util.Optional;
 import java.util.function.LongFunction;
@@ -32,8 +31,7 @@ import java.util.function.LongFunction;
 final class ConfigurableRustIter<E> extends AbstractNativeProxy implements RustIter<E> {
 
   private final LongFunction<E> nextFunction;
-  private final View collectionView;
-  private final ViewModificationCounter modificationCounter;
+  private final ModificationCounter modificationCounter;
   private final Integer initialModCount;
 
   /**
@@ -41,18 +39,15 @@ final class ConfigurableRustIter<E> extends AbstractNativeProxy implements RustI
    *
    * @param nativeHandle nativeHandle of this iterator
    * @param nextFunction a function to call to get the next item
-   * @param collectionView a database view of the collection over which to iterate
    * @param modificationCounter a view modification counter
    */
   ConfigurableRustIter(NativeHandle nativeHandle,
                        LongFunction<E> nextFunction,
-                       View collectionView,
-                       ViewModificationCounter modificationCounter) {
+                       ModificationCounter modificationCounter) {
     super(nativeHandle);
     this.nextFunction = nextFunction;
-    this.collectionView = collectionView;
     this.modificationCounter = modificationCounter;
-    this.initialModCount = modificationCounter.getModificationCount(collectionView);
+    this.initialModCount = modificationCounter.getCurrentValue();
   }
 
   @Override
@@ -62,9 +57,9 @@ final class ConfigurableRustIter<E> extends AbstractNativeProxy implements RustI
   }
 
   private void checkNotModified() {
-    if (modificationCounter.isModifiedSince(collectionView, initialModCount)) {
-      throw new ConcurrentModificationException("Fork was modified during iteration: "
-          + collectionView);
+    if (modificationCounter.isModifiedSince(initialModCount)) {
+      throw new ConcurrentModificationException("Collection or the corresponding Fork "
+          + "were modified during iteration");
     }
   }
 }